Add script that generates a first draft of the release notes
Example:

    python scripts/releasenotes.py 9.6.0 <github_token>

This will output a NEWS-section based on the pull requests and issues
related to the version to be released.

Full disclosure, this script has been generated using an AI assistant as
I couldn't be bothered to learn how to interface with the GitHub REST
API. It seemingly produces a useful result so thanks to our AI
overloads.